% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_mldm.R
\name{biplot.mldm}
\alias{biplot.mldm}
\title{Biplot for fitted MLD model.}
\usage{
\method{biplot}{mldm}(object, myX = c(-4, 4), myY = c(-4, 4),
  resp.var.labels = NULL, ...)
}
\arguments{
\item{object}{An object of fitted model.}

\item{myX}{Range for X-axis.}

\item{myY}{Range for Y-axis.}

\item{resp.var.labels}{Labels of the reponse variable.}
}
\description{
Biplot for fitted MLD model.
}
