% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldm_fit.R
\name{mldm.fit}
\alias{mldm.fit}
\title{Fits a Multivariate Logistic Distance (MLD) model on multivariate binary data (i.e., multiple binary responses with one or many independent explanatory variables).}
\usage{
mldm.fit(formula, index, resp.dim.ind, resp.var.labels = NULL, data = NULL,
  id, center = TRUE, scale = FALSE, Rusr = "independence",
  conf.interval = 0.95, bootstrap = FALSE)
}
\arguments{
\item{formula}{A glm formula, should be specified for each dimension.}

\item{index}{An identifier of the response variable.}

\item{resp.dim.ind}{Response Indicator matrix. It indicates to which dimension 
every response belongs.}

\item{resp.var.labels}{Labels of the reponse variable.}

\item{data}{Input data frame.}

\item{id}{Subject identification number.}

\item{center}{Apply centering on explanatory variables.}

\item{scale}{Apply standard normal transformation on explanatory variables.}

\item{Rusr}{A (working) correlation coeffiecient.}

\item{conf.interval}{Confidence interval level.}

\item{bootstrap}{Apply Clustered Bootrstrap procedure to obtain model standard errors.}
}
\description{
Fits a Multivariate Logistic Distance (MLD) model on multivariate binary data (i.e., multiple binary responses with one or many independent explanatory variables).
}
\examples{
## Load package
# library(mldm)

## Load dataset
# data(NESDA)

## Specifiy response indicator matrix 
# Z <- matrix(c(1,1,0,0,0,0,0,1,1,1), 5, 2, byrow = FALSE)

## Specify model formula
# mf <- Outcome | Outcome ~ EDU + GEN + AGE + N + E | EDU + GEN + AGE + N + E
# mf <- Formula(mf)

## Fit MLD model
# fit1 <- mldm.fit(formula = mf, index = Index, resp.dim.ind = Z, 
#                  data = NESDA, id = pident, scale = TRUE)
                   
## display result
# fit1


}
\author{
Hailemichael M. Worku (aka, Haile) and Mark de Rooij.
2018 | Leiden university, M&S unit.
}
