% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mldm.R
\name{summary.mldm}
\alias{summary.mldm}
\title{Summary of model fit.}
\usage{
\method{summary}{mldm}(object, bootstrap = FALSE, boot.nonparam = FALSE,
  ...)
}
\arguments{
\item{object}{An object of fitted model.}

\item{bootstrap}{Boolean value (TRUE/FALSE) to indicate a Clustered Bootstrap procedure is used for obtaining model standard errors.}

\item{boot.nonparam}{Boolean value (TRUE/FALSE) to ask for non-parametric confidence intervals.}
}
\description{
Summary of model fit.
}
